multiply: check for overflow when multiplying two integers
TEST: verify that a run error occurs if the multiplication of the two
integers results in a value that can't be stored in an integer.

The execute multiply integer-integer function was modified to put the
left hand side value into a 64-bit integer variable and then multiply it
by the right hand side value.  The result is then compared against the
maximum (32-bit) value, and if higher, throws an overflow error.