multiply: implement double-double operator execute function
TEST: verify that the multiply double-double operator gives the correct
result.

To make this test pass, the multiply double-double function was
implemented, which pops the right hand side value from the stack, gets
the left hand side value from the stack, multiplies it by the right hand
side value and places the result in the double value on top of the
stack.  The pointer to this function replaced the null pointer of the
double-double multiply code instance.