multiply: check for overflow when multiplying two doubles
TEST: verify that a run error occurs if the multiplication of the two
doubles result in a value that can't be stored in a double.

The execute multiply double-double function was modified to compare the
result with the maximum double value and throw an error if higher.  Note
that it is not necessary to check for an underflow since doubles will
go to zero in this situation.