multiply: check for overflow when multiplying double-integer
TEST: verify that a run error occurs if the multiplication of a double
and an integer results in a value that can't be stored in a double.

The execute multiply double-integer function was modified to compare the
result with the maximum double value and throw an error if higher.  The
resulting code was nearly identical to the execute multiply
double-double, so the common code was moved into a new in line do double
multiply function.