precedence: changed handling and how precedences are assigned
Up to now, each operator was assigned its own precedence level.
However, some operators are at the same precedence level.  The next
operator to implement, division, is the same as multiplication.
Therefore, the code need to be modified to support multiple operators at
the same precedence level.

Another issue is a duplication of operator keywords.  With single
character operators, this isn't too big of a problem.  However, some
operators consist of multiple characters (for example, the comparison
operators <=, >= and <>; MOD; and the various logic operators AND, OR,
XOR, etc.).  A solution is needed so that these keywords do not need to
be specified in more than one location in the code.

The precedence info class already contains the keywords for the
operators, so this is the logically location for these keywords.  Up to
now, these were only used for recreation, so looking them up be code
value was all that was necessary.  However, the expression compiler
essentially contained another copy of these keywords.  For the
expression compiler to make use of the precedence info, it needed a way
to get these keywords by a method other than by code value (which isn't
available until after the operator is identified).

A precedence level enumeration was added to the precedence class.  These
enumerators are the identifier to look up the operator keywords by.  The
order of these enumerators also now define the precedence levels
(previously this was done by the order of the operator data initializer
list).  A precedence level member has been added to the operator data
structure, which is now used to set the precedence (the current
precedence member of the precedence info class was removed).

Previously the unsigned type was used for the precedence level (though
there was one instance of an integer).  This type was changed to the
new precedence level enumeration type (included the precedence member
of the recreator stack structure).