precedence: made use of precedence class during compiling
The expression compiler functions were modified to make used of the
precedence data, specifically the operator keywords.

A new map was added to the precedence info class that maps the
precedence level to the operator data used to initialize the other maps.
A regular map was used to that the operator data is kept in sorted order
(this will be important later when there a multiple operators at the
same precedence level).  Since this map is keeping a pointer to the
operator data, the operator data initializer list constructor was made
static (a local variable would disappear).

New access functions were added to return the operator codes for a given
precedence, one that just returns the codes (used by the compile
negation function), and one that also takes an operator character.  This
function returns the codes for the precedence only if the character
passed matches that of the operator at that precedence, otherwise it
returns a null pointer.  This function is used by the compile product
and exponential functions.

The is operator character function was changed to the operator codes
function and modified to take a precedence instead of an operator
character.  It now passes the peek next character to the precedence
operator codes function, and returns a pointer to the codes instance or
a null pointer if no codes were return.  The compile negation was
modified to call the precedence operator codes (without the character
argument).

A virtual select function was added to the operator codes base class.
This is so the expression compiler functions can generically call this
function for the codes function returned by the precedence class.  This
function takes two data type arguments.  The unary operator select
function was modified to take an unused dummy data type so that there is
a common select function interface.  This argument was made optional
for base class select function.
The compile product and exponential functions were modified