divide: adding parsing of the divide operator
TEST: verify that a simply divide expression is parsed to the end of the
expression.

The divide operator is at the same precedence as the multiply operator
and so will be compiled by the compile product function.

The four divide code instances were added (with all null pointers) along
with the divide numeric operator codes instance.  An entry for the
divide codes and divide operator character was added to the operator
data list in the precedence info constructor with the same product
precedence level.

So that multiple entries can be added to the operator data map, it was
changed from a standard map to a standard multi-map.  The standard
multi-map does not support the bracket operator, so the emplace function
needed to be used to add to the map.

These relatively minor changes bring quite a bit of functionality
including all divide codes being supported being compiled into code
(because the divide codes include all four - required to compile), and
multiple divide operators (because it is handled in the compile product
function that has already been fully tested).  Therefore, no additional
divide compiling tests will be needed.  Checking if the proper divide
codes are compiled will be covered in the recreate and execute tests.