divide: implement divide of two integers
TEST: verify that division of two integers is executed correctly.

The execute divide integer-integer function was implemented and its
function pointer was put into the correct code instance.  It is not
possible for any kind of overflow or underflow to occur when dividing
one integer by another, so these changes are not necessary.

OTHER: rearranged the operators source file where the code and codes
instances are put after the execute functions (which allowed for the
removal of the function declarations).