multiply: check for an underflow
TEST: verify that a run error occurs if there is an underflow when
multiplying.

It was previously thought that an underflow couldn't occur because the
values would go to zero.  This was incorrect as an underflow is a
negative value that is to large.  Therefore a check was added to the
common multiply and check result function, which is used by the execute
double-double, integer-double and double-integer functions, so no
additional tests are needed for the other operators.