divide: check for an underflow when dividing two doubles
TEST: verify that a run error occurs if there is an underflow when
dividing two doubles.

A check was first added to the execute divide double-double function.

The duplicate overflow checks were now in several places, so common
overflow functions were added (in a new header file), one for a 64-bit
integer to check against the maximum/minimum 32-bit integer, and the
other for checking double overflow.  Both functions are inline and were
also given a reference to the executer argument.  The multiple functions
were modified to use these also.

The divide by zero check was also duplicated (almost) but were different
in the type of the variable being checked.  A template function was
added to cover both an integer and double right hand side value.