exponential: passed executer reference to power classes
Inspired by passing the executer reference to the overflow functions,
the power integer-integer and double-integer function operator classes
were modified to take a reference to the executer so that it is
available when an error is thrown (for the current offset).

This simplifies the executer exponential functions that use these
classes where the extra try/catch block could be removed since it is no
longer necessary to catch the error to add the current offset to it.
This should have little impact on execution since these functions are
defined totally with inline functions.