divide: added overflow check to integer-double divide
TEST: verify that a run error occurs if the division of an integer by a
double results in a value that can't be stored in a double.

The execute divide integer-double function was first modified to call
the check overflow function.  Upon seeing the duplicated code with the
execute divide double-double function, a new divide and check result
inline function was added the does the division, checks for overflow and
sets the top value on the stack to the result.  Both execute functions
were changed to call the new function.  Since this check also includes
the underflow check, no test is needed for that for the integer-double
divide operator.