divide: added divide by zero check to double-integer divide
TEST: verify that a divide by zero run error occurs when attempting to
divide a double by an zero integer.

A check for the right hand side equal to zero was first added by calling
the check divide by zero function.  The duplicated code with the execute
divide integer-integer function was eliminating by adding a new pop
integer divisor function that gets the right hand side value, checks for a
divide by zero, pops the value, and returns the value.  Both execute
functions were changed to call the new function.

There is no need to add an overflow check to this function since there
is no way for an overflow to occur when dividing by the limited range of
an integer (the result will always be less than the dividend).