divide: adding parsing of the integer divide operator
TEST: verify that a simply integer divide expression is parsed to the
end of the expression.

The integer divide operator (back slash character) is one precedence
level above product (multiplication and regular division), so a new
precedence level was added for it.  It is special operator whose purpose
is to divide two doubles and return an integer value, so there is only
one form of the operator (that taking two double operands).

A new integer divide operator code class based on the operators codes
class was added for it.  This class only contains a single code, and
this operator will be the only operator of this codes class.  It was
given the same select function, but it does not use the data type
arguments.

An integer divide code instance was added (with null pointers) and an
instance of integer divide operator code code with this single code.

A new compile integer division function was added to the expression
compiler.  It calls the compile product function, calls the operator
codes for the integer divide precedence level, which if returns a non
null pointer, calls the compile product function again for the second
operand (which was all that was needed to make the test pass).  The
compile numeric expression function was changed to call this new
function.

The divide operator is at the same precedence as the multiply operator
and so will be compiled by the compile product function.

The four divide code instances were added (with all null pointers) along
with the divide numeric operator codes instance.  An entry for the
divide codes and divide operator character was added to the operator
data list in the precedence info constructor with the same product
precedence level.

So that multiple entries can be added to the operator data map, it was
changed from a standard map to a standard multi-map.  The standard
multi-map does not support the bracket operator, so the emplace function
needed to be used to add to the map.

These relatively minor changes bring quite a bit of functionality
including all divide codes being supported being compiled into code
(because the divide codes include all four - required to compile), and
multiple divide operators (because it is handled in the compile product
function that has already been fully tested).  Therefore, no additional
divide compiling tests will be needed.  Checking if the proper divide
codes are compiled will be covered in the recreate and execute tests.