int-divide: added operator code to program code
TEST: verify that with two double operands, that an integer divide
operator code is added after the operand codes.

To get this test pass, functionality was added to the compile integer
division function that mirrors the other functions including getting the
codes (in this case containing just one), getting the data type for the
right hand side operand, calling the select function and adding the
instruction code to the program.

(Yes there is a pattern of duplicate code forming between these compile
functions that will refactored once the full pattern shows itself.)