int-divide: implemented execution of operator
TEST: verify that integer division of two doubles is executed correctly.

The execute integer divide function was implemented and its function
pointer was put into the correct code instance.  This function pops the
right hand side double operand value from the stack, gets the left hand
side double operand value, divides them, and then converts the result to
an integer and puts it on top of the stack.