int-divide: check for underflow when dividing
TEST: verify that an overflow run error occurs if the integer division
of the two double operands result in a value that can't be stored in an
integer.

The execute integer divide function was modified to also compare the
result with the minimum integer value and throw an error if lower.

This overflow check is was similar to one of the check overflow
functions already implemented, but was different in the type of the
result argument (double vs. 64-bit integer).  This function couldn't
immediately be changed to a template with the type of the result was
the template argument because if conflicted with the other check
overflow that checks for overflow of a double.  Therefore, the function
checking for an integer overflow was renamed to check integer overflow
(and was changed to a template), and the other was renamed to the check
double overflow.  This also makes it clearer was is being checked.