int-divide: change left side integer constant to a double constant
TEST: verify when there is an integer on the left side of the
integer divide operator, that the integer constant is changed to a
double constant.

Since constants contain both their integer and double values in the
dictionary, either can be used as needed.  This save time by avoiding an
unnecessary conversion during program execution.  Even though a constant
was interpreted as an integer constant (no decimal point or exponent and
within the range of an integer), its value was converted to a double and
stored in the dictionary.  All that needs to be done is that the
constant integer code put into the program needs to be changed to a
constant double code.

This test initially failed because the execute integer divide function
assumed that the left hand side operand had pushed a double value on top
of the value stack, but for an integer constant, the integer value was
set, and the value was not interpreted correctly giving the wrong
result.

The compile integer division function in the expression compiler class
was changed to call a new convert to double function in the compiler
class.  To make this test pass, all this new function needed to do was
to get the offset where the constant code was added to the code.  The
constant integer was just added, so its code and operand are the last
two words in the code line, therefore its offset is two less than the
size of the code line.  The word at this offset is set to the constant
double code.  If the code is already a constant double, this still
works.