int-divide: change right side integer constant to a double constant
TEST: verify when there is an integer on the right side of the
integer divide operator, that the integer constant is changed to a
double constant.

This test initially failed because the execute integer divide function
assumed that the right hand side operand had pushed a double value on
top of the value stack, but for an integer constant, the integer value
was set, and the value was not interpreted correctly giving the wrong
result.

The compile integer division function in the expression compiler class
was changed to call a new convert to double function in the compiler
class after compiling the right side operand.