int-divide: only change to double constant if operand was a constant
TEST: verify that a non-constant code is not erroneously changed to a
double constant with an integer divide operator.

If the code of the previous operand is not a constant, code will be
corrupted.  This is worsened if the last operand was an operator as the
convert to double could be changing an operand to constant double code.

This test initially failed because the last operand was a divide
operator (of two constants) and when the last code was changed to a
constant double code, if was actually the operand word of the second
constant that was actually changed, and the execution of the code
malfunctioned (in this case, used the wrong constant).

To get this test to pass, a last constant data type member was added to
the compiler class.  For a constant, this is set to the data type of a
constant in the add numeric constant instruction function, but changed
to a null data type in the add instruction function.  The covert to
double function was changed to check if the last constant data type is
integer before changing to a constant double code.  This change also
now prevents changing a constant double if it is already a constant
double.

OTHER: the add constant numeric instruction function was renamed to add
numeric constant instruction as it reads better.