int-divide: add convert to double code if operand is integer
TEST: verify that the internal convert to double code is added to the
code if for a non-constant integer operand.

The special convert to double to double code will be inserted into the
program for integer divide operator and later for functions meant to
take a only so that many codes with don't need to be implemented.  The
exception being the math operators that that could be used with either
type (otherwise these convert to double, and integer, codes would be in
code in many places, which could effect execution time, and requiring
the user to use the CVTDBL and CVTINT functions to avoid this is an
undesirable burden).

To get this test to pass required a number of changes.  A convert to
double code instance (with null pointers for the recreate
and execute functions) was added.  The last constant data type was
changed to the last operand was constant boolean flag, set to false
for non-constants and true for constants.  The convert to double
function was given an operand data type argument, which is now checked
to see if a conversion to double is needed (the operand data type was
added to the two calls).  If the last operand was not a constant (the
new flag is false), then the internal convert to double code is added to
the program.

OTHER: missed committing the constant number unit tests source file for
the renaming of the add numeric constant instruction function.