int-divide: recreate convert to double code to nothing
TEST: verify that an internal convert to double code is recreated to
nothing.

To make this test pass, a recreate nothing function was implemented that
does nothing, and a pointer to this function replaces the null pointer
in the convert to double code instance.

OTHER: the integer divide conversion convert to double test code was
restructured a bit (to that some reuse of setup code can be utilized
for the execute test).