int-divide: execute the internal convert to double code
TEST: verify that an expression requiring an internal convert to double
code is executed correctly.

An execute convert double function was implemented that converts the
integer on top of the stack to a double and puts it into the double
value on top of the stack.  A pointer to this function replaces the null
pointer in the convert to double code instance.