int-divide: compile multiple integer divide operators
TEST: verify that multiple integer divide operators are compiled.

This test initially failed because the compile integer division function
only looked for one integer divide operator (because it only contained
an if statement).  Even though possible, multiple integer divide
operators wouldn't normally be used (since the result is integer, the
result would need to be converted to a double for use as an operand of
another integer divide operator).

To make this test pass the if was changed to a while statement.