modulo: parse the mod word operator
TEST: verify that the mod operator is parsed along with the operand
following it.

This is the first operator to be implemented that consists of a word
instead of a symbol, so simply looking at the next character is not
sufficient.

To make this test pass, a compile modulo function was added and the
compile numeric expression was changed to call it instead of the compile
integer division function.  The compile modulo function in turn calls
the integer division function.  A new word operator codes function was
added to mirror the operator codes function (renamed to symbol operator
codes to parallel the new word function).  To get this function to skip
over the word, the compile get keyword function was called.

Since the test only checks that the MOD keyword is parsed, this was
sufficient.  Following this, the compile modulo function calls the
compile integer division function again to get the second operand.

This did make the new modulo unit tests (parse the MOD operator and its
two operands), the changes caused one of the exponential unit tests to
fail - the "3a2" test that was suppose to stop at the 'a' because it is
invalid, no longer stops because 'a' is now considered a word and is
read.

To get this test to pass again, a word cache was added to the compiler
class.  The get keyword function, after reading the keyword stores the
word in the word cache.  The peek next character function was modified
to return the first character of the word if the word cache is not
empty.  A clear word function was added to clear the word cache.

The command compiler, after reading the command keyword, a call to the
clear word function was added (indicating the word has been accepted).
The new word operator codes function was then modified to check if the
keyword is the MOD keyword, and if it is, the word cache is cleared and
true is return, otherwise false is returned.  This boolean return value
is temporary (this function will return codes for the operator, but the
MOD codes haven't been added yet, so this is only being used to flash
out the word caching mechanism).  The compile modulo was then changed
to only call the compile integer division a second time if the word
operator codes returns true.