modulo: implemented compiling of all of the mod operators
TEST: verify that all of the mod operators are compiled correctly.

To speed development along, this test has a single expression that
tests all four of the modulo operators (plus throws in different cases
of the MOD keyword).

To make this test compile and pass, the four mod operator code instances
was added along with the mod codes instance.  A module precedence level
was added along with adding an entry for the mod operator to the
operator data list in the precedence info constructor.

A new operator codes function was added to the precedence info class to
look up a code by a word instead of a single character.  Except for
comparing a string, the function is the same as the existing operator
codes function.  A static pass through function was added to the
precedence class.

The word operator codes function in the expression compiler class was
modified to its final form where it calls the operator codes taking a
case insensitive word string in the precedence class, clearing the word
in the compile and return the codes if a non-null codes pointer is
returned, otherwise returning a null pointer.

The final form of the compile modulo function was also implement, which
mirrors the other functions except the word operator codes function is
called instead of the symbol operator codes function.  This code has
essentially be tested before (it is a copy, though bad, this will be
refactored later once all the operators have been implemented).