modulo: implemented recreation of the divide operator
TESTS: verify that all four mod operators are recreated correctly.

To speed along development, and since the changes are minor (replacing
the null pointer with a pointer to the recreate binary operator
function), the recreation of all four mod operators was implemented.
(The tests consists of various spaces and no spaces and casing.)