module: added divide by zero check to integer-integer mod operator
TEST: verify that a divide by zero run error occurs when attempting to
mod with a zero.

Since the modulo operation is really a division but returns the
remainder instead of the quotient, the possibly of a divide by zero can
occur, so the right hand side value needs to be checked for a zero
value.  The pop integer divisor function does what is needed, however,
it was discovered that this function had the wrong return type, so it
was changed from the incorrect double to the correct integer (this did
not affect any of the unit tests).

The test only checks if an error is thrown.  This is an issue with
checking for the correct error message because currently the error
length is fixed at one and with the mod operator, the error length needs
to be three characters, which is more than should be changed in this
commit.