recreator: added support for multiple character errors
TEST: verify when a run error occurs at a mod operator that all three
characters of the operator are highlighted.

This test initially failed because the error length was fixed at one
character (all operators up to this time have been one character).

To make this test pass, an end error marker was added.  When the binary
operator is appended, the end of the operator is now also marked.  When
the program error is constructed, the length is now calculated as the
difference between the start error marker character and the end error
marker character (minus one).  The erase of the end marker character was
also added.  The length member of the program error class was changed
from an unsigned integer to the size type.