module: added divide by zero check to double-double mod operator
TEST: verify that a divide by zero run error occurs when attempting to
mod with a zero.

The pop double divisor function does the popping of the right hand side
value and checks for a divide by zero error, to a call to this function
was put into the execute modulo double double function replacing the
existing code to get and pop the right hand side value.