modulo: added divide by zero check to double-integer mod operator
TEST: verify that a divide by zero run error occurs when attempting to
mod a double by a zero integer.

Since the pop integer divisor does the divide by zero check, the get and
pop right hand side statements were replaced with a call to this
function.

OTHER: Since there were frequent getting the top integer value and
converting it to a double (via a static cast), a new executer top
integer as double function was added to do this to remove this
duplication.