add: added parsing of the add operator
TEST: verify that a simply integer add expression is parsed to the end
of the expression.

The add operator (plus character) is one precedence level above mod
operator and shares precedence with the subtract operator, so a new
summation precedence level was added for it.

The four add code instances were added (with null pointers) and an
instance of add numeric operator codes with these code instances was
added.

A new compile summation function was added to the expression compiler.
It calls the compile modulo function, calls the symbol operator codes
for the summation precedence level, which if returns a non-null pointer,
calls the compile modulo function again for the second operand (which
was all that was needed to make theis test pass).  The compile numeric
expression function was changed to call this new function.