add: implemented compiling of all of the add operators
TEST: verify that all of the add operators are compiled correctly.

To speed development along, this test has a single expression that
tests all four of the add operators.

The final form of the compile summation function was also implemented,
which mirrors the other functions.  This code has essentially be tested
before (it was copied, and though bad, will be refactored later once all
the operators have been implemented).