add: check for overflow when adding two integers
TEST: verify that a run error occurs if the addition of the two integers
results in a value that can't be stored in an integer.

The execute add integer-integer function was modified to put the left
hand side value into a 64-bit integer variable and then add to it the
right hand side value.  The result is then compared against the maximum
(32-bit) value, and if higher, throws an overflow error.