add: check for underflow when adding two integers
TEST: verify that a run error occurs if the addition of the two integers
results in a value that can't be stored in an integer.

The execute add integer-integer function was modified to also check if
the result is not less than the minimum value that can be stored in an
integer.  (This is still reported as an 'overflow' error.)  Seeing that
this code is the same as the check integer overflow function, the code
was replaced with a call to this function.