negate: added check for underflow with negate integer
TEST: verify that an integer negate of the minimum integer value
produces and overflow run error.

The minimum integer value, cannot be negated as its positive value is
one above the maximum integer value, therefore this is an underflow
(reported as an overflow).

A check was added to the execute negate integer function if the operand
is equal to the minimum integer value, then an overflow error is thrown
(this was better then negating into a 64-bit integer and then checking
for an overflow, and this is the only number that can cause this error).

This was the first error produced by a unary operator, so the recreate
unary operator function needed to be modified to mark the start and end
characters around the unary operator.  To clean up a big, this code was
put into a new append unary operator function (which mirrors the
recreate binary operator functions).