corrected bug with double overflow check
TEST: verify that a negative value does not cause an overflow error.

A problem was discovered with the double overflow check.  Unlike with
integers where the minimum numeric limit is the lowest possible integer
value (so an overflow, or 'underflow' is the condition of a negative
number being too high), the minimum numeric limit for a double is the
'smallest' possible double value, which is essentially equivalent to
zero, so an double 'underflow' isn't a condition that needs to be
checked because it is zero.  However, what does need to be checked for
is a negative value that is too high.

It just so happens with the unit tests implemented so far, that checking
a double value for less than the minimum value allowed the tests to
pass.  But, for a multiply test for a simple negative number result, an
overflow error occurred because any negative number will be less then
the smallest double value (essentially 0).

To make this test pass and all the previous tests continue to pass, the
check double overflow function was modified where the test for less than
the minimum double value was removed and the test for higher than the
maximum double value changed where the absolute value of the result is
compared - this will also catch negative values that 'overflow' what can
be stored in a double.