subtract: added parsing of the subtract operator
TEST: verify that a simple integer subtract expression is parsed to the
end of the expression.

The subtract operator (minus character) is the same precedence level add
the add operator.

The four subtract code instances were added (with null pointers) and an
instance of subtract numeric operator codes with these code instances
was added.  An entry was added to the precedence list for the subtract
codes.

No further compile tests are needed since the subtract operator uses the
same already tested compile summation function.