subtract: check for underflow when subtracting two integers
TEST: verify that a run error occurs if the subtraction of the two
integers results in a value that can't be stored in an integer.

The execute subtract integer-integer function was modified to also check
if the result is not less than the minimum value that can be stored in
an integer.  (This is still reported as an 'overflow' error.)  Seeing
that this code is the same as the check integer overflow function, the
code was replaced with a call to this function.