precedence: refactored how operator precedence data is initialized
The original method for defining the operator precedences was the order
that the operators were listed in the operator data list in the
precedence info class constructor.  This was prior to the addition of
the precedence level to support operators at the same precedence, which
is now how the precedence levels are set, and the order the operators
are listed is not important.

Since there is a one to one relation between each codes instances and
each entry in the operator data list, it made more sense that the
precedence level and the operator keyword be defined in the code
instances.  The code was modified with a couple of refactoring steps.

The initialize operator data function was changed to take the individual
fields of the operator data instead of a reference to the operator data
structure.  This will be needed when called from the operator codes
class constructors for entering this data from the codes instances.

Added a static add operator codes function to the precedence class with
precedence, codes instance reference and keyword string.  This function
first gets the precedence info instance and then calls the initialize
operator data function (which was renamed to add operator codes and made
public), passing the arguments through.

Precedence level and keyword arguments were added to the unary, numeric
and integer divide operator codes constructors.  Calls to the new static
add operator codes were added to these constructors passing the new
arguments through with a reference to the this instance.  The operator
data (precedence and keyword) from the precedence info constructor was
moved to the codes instances.  The precedence info constructor body was
removed (this constructor is still needed and it still private).

Finally, the precedence member was removed from the private operator
data structure since it was only used by the precedence level
constructor.