less-than: adding parsing of the less than operator
TEST: verify that a simple less than expression is parsed to the end of
the expression.

The less than operator is at a new relation precedence that will be
shared with the three other relational operators.

Two of the four relational operators consist of two characters so the
current get symbol operator codes function would not work.  Compounding
the problem is the first character of the less than and less than or
equal operators is the same as the not equal operator, which is at a
different lower precedence shared with the equal operator.

Therefore, a new get comparison operator codes function was added to the
compiler class.  This function will handle all of the comparison
operators including the equality operators.  For now, this function
skips white, creates a keyword of the next character, calls a new
comparison operator data static function in the precedence class, and if
codes are returned, the next character is consumed and the codes are
returned.

The static comparison operator data function gets the precedence info
instance and calls its comparison operator data.  This function uses a
new map from keyword to operator codes pointer.  This map is searched
for the keyword.  If the keyword is not found then a null pointer is
returned, otherwise the codes are returned.  In the add operator data
function, if the precedence is relation, then the codes of the operator
is added to this new map by keyword.

The four less than code instances were added (with all null pointers)
along with the less than numeric operator codes instance with their
precedence level and keyword.

A new compile relation function was added to the expression compiler
that calls the compile summation function, calls the new get comparison
operator codes, and if codes are returned calls the compile summation
function again for the second argument, and returns the data type.  The
compile numeric expression was changed to call the compile relation
function.

OTHER: the operator data map member was renamed removing the unnecessary
'map' suffix.