less-than: implemented compiling of all of the less than operators
TEST: verify that a less than expression is compiled correctly.

To get this test to pass, the compile relation function was modified
to contain similar code as the other compile function including getting
the data type from the right hand side operand, selecting the
appropriate code for the operand data types and adding the less than
instruction to the program code.

Only the double-double less than operator was tested, but these changes
support all four less than codes (this code has been tested before, so
additional test are not necessary, but will be covered with the
recreation and execution tests of each of less than codes).