less-than: implemented comparing of two doubles
TEST: verify that less than of two doubles is executed correctly.

The execute less than double double function was implemented, which pops
the right hand side value from the stack, compares it to the left hand
side value, and sets the top integer on the stack to true if the left
hand side value is less than the right hand side value.  In BASIC, a
true value is all bits of the integer on, which is the same as -1, and
false is all bits off, which is zero.

In realizing that this conversion of true to -1 and false to 0 is going
to be repeated for every operator for all data types, a new set top from
boolean value function was added to the executer class, which takes a
boolean value and converts it to an integer of -1 or 0.

The test evaluates less than with three expressions, one that returns
true, one that returns, and one with both value equal (which should also
return false for less-than).  This covers both branches of the tertiary
operator when setting the top integer value from both boolean states.

Seeing that having a set function for this was desirable, the pattern
was continued by replacing the top function (which returns a reference
to the top item on the stack) with various access functions including
top double, top integer, set top double and set top integer.  There was
already a top integer as double function that follows this pattern.

The top function was then removed.  Returning a reference to an internal
member broke encapsulation anyway and was good to have it replaced.